copilot-c99: Redefine C99 translation of signum affecting 0, NaN. Refs #278.
The semantics of the Copilot stream function signum is inconsistent
between the C99 backend and the interpreter.

`copilot-c99` translates `signum e` as `copysign(1.0, e)`, which is not
the same function for some corner cases.

For example, as evaluated by the interpreter, `signum 0` returns `0`,
but the C99 code generated for it, `copysign(1.0, 0)`, returns `1`.

Moreover, when dealing with floating-point numbers, `signum (-0)` and
`signum NaN` return `-0` and `NaN`, respectively, in the interpreter.

The `copysign`-based implementation returns `-1.0` and `±1.0` (depending
on the sign bit of the `NaN`), respectively.

This commit replaces the `copysign`-based implementation such that
`signum e` now returns `e > 0 ? 1 : (e < 0 ? -1 : e)`. That is, if `e >
0`, return `1`, else if `e < 0`, return `-1`, else return `e`.

In particular, this affects the cases when `e` is `±0` or `NaN`, which
now fall back to the "return `e`" case. Note that this changes the API
of the function: it previously returned a value of -1 or 1, and now can
also return -0, 0, -NaN or NaN.